remove implicit dep between mysql and redis typing
