+++ pkg/devtools/app/src/components/pages/MySQL.vue
@@ -439,7 +439,7 @@ import { useMySQL, getStatementType } from '@/composables/mysql'
 import { formatTimestamp, formatBytes, formatDuration } from '@/composables/formatters'
 import { useColumnResize, type ColumnConfig } from '@/composables/columnResize'
 import { useMySQLStore } from '@/stores/mysql'
-import type { DatabaseRequest } from '@/stores/redis'
+import type { DatabaseRequest } from '@/types/database'
 import DirectionIndicator from '@/components/ux/DirectionIndicator.vue'
 import Toolbar from '@/components/ux/Toolbar.vue'
 import ColumnToggle from '@/components/ux/ColumnToggle.vue'
