+++ pkg/devtools/app/src/components/pages/Redis.vue
@@ -453,7 +453,7 @@ import { useRedis, getCommandFromStatement } from '@/composables/redis'
 import { formatTimestamp, formatBytes, formatDuration } from '@/composables/formatters'
 import { useColumnResize, type ColumnConfig } from '@/composables/columnResize'
 import { useRedisStore } from '@/stores/redis'
-import type { DatabaseRequest } from '@/stores/redis'
+import type { DatabaseRequest } from '@/types/database'
 import DirectionIndicator from '@/components/ux/DirectionIndicator.vue'
 import Toolbar from '@/components/ux/Toolbar.vue'
 import ColumnToggle from '@/components/ux/ColumnToggle.vue'
