+++ pkg/devtools/app/src/composables/redis.ts
@@ -1,6 +1,7 @@
 import { computed, watchEffect } from 'vue'
 import { storeToRefs } from 'pinia'
-import { useRedisStore, type DatabaseRequest } from '@/stores/redis'
+import { useRedisStore } from '@/stores/redis'
+import type { DatabaseRequest } from '@/types/database'
 import { useUrlParams } from '@/composables/urlParams'
 import { useBufferSettings } from '@/composables/bufferSettings'
 import { usePersistedFilters } from '@/composables/persistedFilters'
